Drop guards the pixel conversion made dead
Kit::dpr_ is a positive invariant: it starts at 1.0f, and both writers
clamp, so the "dpr > 0 ? dpr : 1" dance at every read was checking
something that cannot happen.  thumb_dest_params() keeps its clamp, being
the one that takes dpr as a parameter.

Browser's set_dpr() re-tested a delta that its callee had already decided
on; Splitter::min_w's fallback repeated the same 8.0f its default and its
only assignment already spell out.

GridRow::y/h and layout_w_ held integers in floats.  Kit::px() was
open-coded three times.  kGlow duplicated kGlowPts.  Also removes
kWinPadY, which had no uses before this series either.

+++ dn/kit-browser.hpp
@@ -106,8 +106,8 @@ struct Browser : Widget {
 	struct GridRow {
 		int first = 0;
 		int count = 0;
-		float y = 0;
-		float h = 0;
+		int y = 0;
+		int h = 0;
 	};
 	std::vector<GridRow> rows_;
 	int cursor_ = -1;
@@ -115,7 +115,7 @@ struct Browser : Widget {
 	bool cursor_x_dirty_ = false;
 	int layout_cursor_ = -1;
 	float layout_cell_x_ = 0;
-	float layout_w_ = 0;
+	int layout_w_ = 0;
 	int mid_file_ = -1;
 
 	std::vector<File> files_;
